BOSA21Q1-246: fix n+1 queries problem in the app
Updated:

Initially I have fixed n+1 queries for initiative card view. But I was
talking with Oskar about this and he mentioned about goldiloader gem. In
the result this gem was solved automatically what I have done and
for other places which we don't know. Moreover I decided to keep random
as a default filter option.